# improve graph traversal methods