Similarly to #39216, we avoid the conversion to `short_digraph` when the input graph is an instance of `StaticSparseBackend`. 

Furthermore, we use the new `PairingHeap_of_n_integers` data structure (#39046) instead of a `priority_queue` to emulate a max-heap. This is slightly faster this way and cleaner.


### :memo: Checklist

<!-- Put an `x` in all the boxes that apply. -->

- [x] The title is concise and informative.
- [x] The description explains in detail what this PR is about.
- [x] I have linked a relevant issue or discussion.
- [x] I have created tests covering the changes.
- [ ] I have updated the documentation and checked the documentation preview.

### :hourglass: Dependencies

<!-- List all open PRs that this PR logically depends on. For example, -->
<!-- - #12345: short description why this is a dependency -->
<!-- - #34567: ... -->