@dcoudert 2024-12-28
Some timings to see the (small) advantage of using the pairing heap data structure.

Before
```sage
sage: G = graphs.PathGraph(4)
sage: %timeit G.maximum_cardinality_search(initial_vertex=0)
9.77 µs ± 16.9 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
sage: G = graphs.Grid2dGraph(10, 10)
sage: %timeit G.maximum_cardinality_search()
209 µs ± 1.53 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
sage: G = graphs.Grid2dGraph(100, 100)
sage: %timeit G.maximum_cardinality_search()
87.2 ms ± 462 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
sage: G = graphs.CompleteGraph(10)
sage: %timeit G.maximum_cardinality_search()
27 µs ± 85.9 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
sage: G = graphs.Cell600()
sage: %timeit G.maximum_cardinality_search()
355 µs ± 1.6 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

With this PR:
```sage
sage: G = graphs.PathGraph(4)
sage: %timeit G.maximum_cardinality_search(initial_vertex=0)
9.74 µs ± 32.6 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
sage: G = graphs.Grid2dGraph(10, 10)
sage: %timeit G.maximum_cardinality_search()
206 µs ± 371 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
sage: G = graphs.Grid2dGraph(100, 100)
sage: %timeit G.maximum_cardinality_search()
85.9 ms ± 363 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
sage: G = graphs.CompleteGraph(10)
sage: %timeit G.maximum_cardinality_search()
25.4 µs ± 36.5 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
sage: G = graphs.Cell600()
sage: %timeit G.maximum_cardinality_search()
324 µs ± 1.2 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```